+++ crates/matrix-sdk/CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
+- Allow `Client::get_dm_room()` to be called without the `e2e-encryption` crate feature.
+  ([#5787](https://github.com/matrix-org/matrix-rust-sdk/pull/5787))
 - [**breaking**] Add `encryption::secret_storage::SecretStorageError::ImportError` to indicate
   an error that occurred when importing a secret from secret storage.
   ([#5647](https://github.com/matrix-org/matrix-rust-sdk/pull/5647))